Nano: code cleanup
